Add pandas profiling analysis (optional) in api
